+++ backend/canisters/user_index/impl/src/updates/c2c_notify_events.rs
@@ -133,7 +133,7 @@ fn process_new_user(
     state
         .data
         .users
-        .register(caller, user_id, username.clone(), now, referred_by, UserType::User);
+        .register(caller, user_id, username.clone(), now, referred_by, UserType::User, None);
 
     state.data.local_index_map.add_user(local_user_index_canister_id, user_id);