+++ backend/libraries/types/src/user_summary.rs
@@ -1,4 +1,6 @@
-use crate::{DiamondMembershipDetails, DiamondMembershipStatus, DiamondMembershipStatusFull, SuspensionDetails, UserId};
+use crate::{
+    BotConfig, DiamondMembershipDetails, DiamondMembershipStatus, DiamondMembershipStatusFull, SuspensionDetails, UserId,
+};
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 
@@ -34,6 +36,7 @@ pub struct UserSummaryStable {
     pub suspended: bool,
     pub diamond_membership_status: DiamondMembershipStatus,
     pub is_unique_person: bool,
+    pub bot_config: Option<BotConfig>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]